// catalog service/generated/zod/inputTypeSchemas/ProductAvgOrderByAggregateInputSchema.ts
import type { Prisma } from '../../prisma';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const ProductAvgOrderByAggregateInputSchema: z.ZodType<Prisma.ProductAvgOrderByAggregateInput> = z.object({
  price: z.lazy(() => SortOrderSchema).optional(),
  price2: z.lazy(() => SortOrderSchema).optional(),
  discount: z.lazy(() => SortOrderSchema).optional(),
  stock: z.lazy(() => SortOrderSchema).optional(),
  sold: z.lazy(() => SortOrderSchema).optional(),
  like_count: z.lazy(() => SortOrderSchema).optional(),
  average_rating: z.lazy(() => SortOrderSchema).optional(),
}).strict();

export default ProductAvgOrderByAggregateInputSchema;